+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - fix documentation typo
  - introduce `pretty` log format for `--error-format`, `--debug-format`, `--request-format`, and `--task-format`
   - clippy cleanups: don't borrow references that are immediately dereferenced by the compiler: https://rust-lang.github.io/rust-clippy/master/index.html#needless_borrow
+  - consistently report users simulated, target host(s), start and end times, and total duration of test both in text and metrics and html report
 
 ## 0.13.0 July 19, 2021
   - enable [`gzip`](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html#method.gzip) support and set Accept-Encoding header by default in the client; disable with `--no-gzip` or `GooseDefault::NoGzip`